113.00.00
- Added `Async.Std.Printf` module so that one doesn't unintentionally use
  blocking `Core.Std.Printf` functions in an Async program.

    There was much pre-existing code that did this via:

    : open Core.Std
    : open Async.Std

    `Async.Std.Printf` defines blocking functions (e.g `printf`,
    `eprintf`) to cause a type error, but leaves Async-friendly functions
    (e.g. `sprintf`, `ksprintf`) untouched.

    Replaced uses of `Printf.*`, with `Core.Std.Printf.*` where needed.

+++ test/dump_core_on_job_delay_test_runner.sh
@@ -3,8 +3,9 @@
 set -e -u -o pipefail
 
 rm -f core.* starting-to-block
-if ./dump_core_on_job_delay_test.exe; then
+if output=$(./dump_core_on_job_delay_test.exe 2>&1); then
     rm -f starting-to-block
+    echo "$output"
     echo >&2 "
 Failure: $0
   The program did not appear to be aborted."
@@ -12,14 +13,13 @@ Failure: $0
 else    
     rm -f core.*
     if ! [ -e starting-to-block ]; then
+        echo "$output"
         echo >&2 "\
 Failure: $0
   The program was aborted too soon."
         exit 1
     fi
     rm -f starting-to-block
-    echo "\
-Success: $0
-  It is normal to see the 'Aborted ... (core dumped)' line above." 
+    echo "Success: $0"
 fi
